Definition of DTM and DSM + hillshade
In the introduction of the episode (Section: Raster Projection in R), the definitions of DTM and DSM do not match the ones on the figure. I'm correcting them to be consistent with GIS theory. I'm also adding a short definition of "hillshade". 
This contribution was previously discussed on an issue:  =/=datacarpentry/=/=r-raster-vector-geospatial/issues/331